fixup! create pages/main
