+++ app.js
@@ -1,14 +1,9 @@
 const bulk = require('bulk-require')
 const mount = require('choo/mount')
-const html = require('choo/html')
 const log = require('choo-log')
-const css = require('sheetify')
 const choo = require('choo')
 
 require('./_app')
-
-// // css
-// css('tachyons')
 
 // // init
 // const opts = {}
@@ -26,13 +21,3 @@ require('./_app')
 // // start
 // app.use(log())
 // mount('body', app.start({ href: false }))
-
-// function mainView (state, prev, send) {
-//   return html`
-//     <body>
-//       <main>
-//         <h1 class="f-6">choo choo</h1>
-//       </main>
-//     </body>
-//   `
-// }
